Move communication handlers to dedicated directory
