+++ src/communication/channels/direct.comm-channel.spec.ts
@@ -6,7 +6,8 @@ import { Unit } from '../../unit';
 import { CommError } from '../comm-error';
 import { CommReceiver, CommResponder } from '../comm-handler';
 import { CommPacket } from '../comm-packet';
-import { CommProcessor, createCommProcessor, proxyCommProcessor } from '../comm-processor';
+import { CommProcessor } from '../comm-processor';
+import { createCommProcessor, proxyCommProcessor } from '../handlers';
 import { DirectCommChannel } from './direct.comm-channel';
 
 interface TestPacket extends CommPacket {
