+++ src/communication/handlers/index.ts
@@ -1,0 +1,2 @@
+export * from './create-comm-processor';
+export * from './proxy-comm-processor';
