+++ src/communication/index.ts
@@ -8,4 +8,5 @@ export * from './comm-method';
 export * from './comm-packet';
 export * from './comm-processor';
 export * from './communicator';
+export * from './handlers';
 export * from './methods';
